refactor: move helper functions to GlobalHelperFunc.ts
- Created `GlobalHelperFunc.ts` to house helper functions.
- Moved `BoldTreatment`, `HeaderTextSize`, `PopulateWinnerName`, `RoundOrGame`, and `ToggleGameScreen` functions to `GlobalHelperFunc.ts`.
- Updated imports in `Global.ts`, `PreGame.ts`, `PlayerHandler.ts`, `InGame.ts`, `CompareScores.ts`, `DetermineWinner.ts`, and `Templates.ts` to use functions from `GlobalHelperFunc.ts`.
